added three test functions in the messageview test

+++ app.py
@@ -289,7 +289,7 @@ def messages_add():
 def messages_show(message_id):
     """Show a message."""
 
-    msg = Message.query.get(message_id)
+    msg = Message.query.get_or_404(message_id)
     return render_template('messages/show.html', message=msg)
 
 
@@ -302,9 +302,13 @@ def messages_destroy(message_id):
         return redirect("/")
 
     msg = Message.query.get(message_id)
-    db.session.delete(msg)
-    db.session.commit()
-
+    # prevent deleting other user's warbles
+    if msg.user.id != g.user.id:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
+    else:
+        db.session.delete(msg)
+        db.session.commit()
     return redirect(f"/users/{g.user.id}")
